test: http client product.http

+++ controllers/controller.go
@@ -13,7 +13,7 @@ type pagingResult struct {
 	Limit     int   `json:"limit"`
 	PrevPage  int   `json:"prevPage"`
 	NextPage  int   `json:"nextPage"`
-	Count     int64 `json:"count"`
+	Count     int `json:"count"`
 	TotalPage int   `json:"totalPage"`
 }
 
@@ -27,7 +27,7 @@ func (p *pagination) paginate() *pagingResult {
 	page, _ := strconv.Atoi(p.ctx.DefaultQuery("page", "1"))
 	limit, _ := strconv.Atoi(p.ctx.DefaultQuery("limit", "12"))
 
-	ch := make(chan int64)
+	ch := make(chan int)
 	go p.countRecords(ch)
 
 	offset := (page - 1) * limit
@@ -54,9 +54,9 @@ func (p *pagination) paginate() *pagingResult {
 
 }
 
-func (p *pagination) countRecords(ch chan int64) {
+func (p *pagination) countRecords(ch chan int) {
 	var count int64
 	p.query.Model(p.records).Count(&count)
 
-	ch <- count
+	ch <- int(count)
 }
